feat(database): :card_file_box: reintegration of drizzle
reintegrated drizzle schema. adjusted columns table to use non resevered words. adjusted user table to allow for clerk id

Resolves #5, resolves #15. prepares integration of webhooks for #4 and #7
