+++ project/db_migrations/meta/_journal.json
@@ -5,8 +5,8 @@
     {
       "idx": 0,
       "version": "7",
-      "when": 1753082245580,
-      "tag": "0000_handy_the_hunter",
+      "when": 1753427100141,
+      "tag": "0000_sharp_greymalkin",
       "breakpoints": true
     }
   ]
