+++ project/drizzle.config.ts
@@ -5,7 +5,7 @@ import { defineConfig } from 'drizzle-kit';
 
 export default defineConfig({
   out: './db_migrations',
-  schema: './db/schema.ts',
+  schema: './lib/db/schema.ts',
   dialect: 'postgresql',
   dbCredentials: {
     url: process.env.DATABASE_URL!,
