+++ project/lib/db/index.ts
@@ -37,8 +37,13 @@ export const queries = {
 }
 */
 
-// Placeholder exports to prevent import errors
-export const db = "TODO: Implement database connection"
+
+import {config} from "dotenv";
+import { drizzle } from 'drizzle-orm/neon-http';
+import * as schema from '@/db/schema';
+
+config({path:".env"});
+export const db = drizzle(process.env.DATABASE_URL!,{schema});
 
 export const queries = {
   projects: {
